+++ Database/__init__.py
@@ -1,0 +1,3 @@
+import init_database
+
+init_database.init_db()
